Add Adventure 6
This is based on Adventure 5, with additions by David Long,
plus other additions performed by an anonymous coder around 1984.
It has a maximum score of 551 points.
Doug McDonald took the Fortran source and changed it so that it will
compile with f77 or f2c, and released version 6.6 on Usenet in August 1990.